requirements.txt
Flask
flask-bcrypt
pymongo
python-dotenv
flask-cors
PyJWT
